+++ exporter/opentelemetry-exporter-otlp-pyproto-common/src/opentelemetry/exporter/otlp/pyproto/common/_internal/_log_encoder/__init__.py
@@ -11,10 +11,10 @@ from opentelemetry.exporter.otlp.pyproto.common._internal import (
     _encode_trace_id,
     _encode_value,
 )
-from opentelemetry.pyproto.collector.logs.v1.logs_service_pyproto2 import (
+from opentelemetry.pyproto.collector.logs.v1.logs_service_pypb2 import (
     ExportLogsServiceRequest,
 )
-from opentelemetry.pyproto.logs.v1.logs_pyproto2 import (
+from opentelemetry.pyproto.logs.v1.logs_pypb2 import (
     LogRecord,
     ResourceLogs,
     ScopeLogs,
