+++ opentelemetry-pyproto/src/opentelemetry/pyproto/collector/metrics/v1/metrics_service_pypb2.py
@@ -7,7 +7,7 @@ Field numbers:
 
 from __future__ import annotations
 
-from opentelemetry.pyproto.metrics.v1.metrics_pyproto2 import ResourceMetrics
+from opentelemetry.pyproto.metrics.v1.metrics_pypb2 import ResourceMetrics
 from opentelemetry.pyproto._pyprotobuf.fields import msg
