+++ opentelemetry-pyproto/src/opentelemetry/pyproto/collector/trace/v1/trace_service_pypb2.py
@@ -7,7 +7,7 @@ Field numbers:
 
 from __future__ import annotations
 
-from opentelemetry.pyproto.trace.v1.trace_pyproto2 import ResourceSpans
+from opentelemetry.pyproto.trace.v1.trace_pypb2 import ResourceSpans
 from opentelemetry.pyproto._pyprotobuf.fields import msg
