+++ opentelemetry-pyproto/src/opentelemetry/pyproto/metrics/v1/metrics_pypb2.py
@@ -39,11 +39,11 @@ from __future__ import annotations
 
 from struct import pack
 
-from opentelemetry.pyproto.common.v1.common_pyproto2 import (
+from opentelemetry.pyproto.common.v1.common_pypb2 import (
     InstrumentationScope,
     KeyValue,
 )
-from opentelemetry.pyproto.resource.v1.resource_pyproto2 import Resource
+from opentelemetry.pyproto.resource.v1.resource_pypb2 import Resource
 from opentelemetry.pyproto._pyprotobuf.fields import (
     byt,
     bool_field,
